refactor(utility): moved out regex_search to utility.py
